add bot and menstrual ui
