+++ mindfullcampus/src/components/chat/Chatbot.css
@@ -1,0 +1,51 @@
+.chatbot-container {
+    width: 400px;
+    background-color: #ffffff;
+    border-radius: 10px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    padding: 20px;
+    margin: 0 auto;
+  }
+  
+  .chatbot-title {
+    background-color: white;
+    font-size: 20px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+  
+  .chatbot-messages {
+    max-height: 300px;
+    overflow-y: auto;
+    margin-bottom: 20px;
+  }
+  
+  .chatbot-message {
+    padding: 10px;
+    border-radius: 5px;
+    margin-bottom: 10px;
+  }
+  
+  .chatbot-user {
+    background-color: #1e90ff;
+    color: white;
+    align-self: flex-end;
+  }
+  
+  .chatbot-bot {
+    background-color: #f5f5f5;
+    color: black;
+    align-self: flex-start;
+  }
+  
+  .chatbot-input {
+    border: none;
+    border-radius: 5px;
+    padding: 10px;
+    margin-bottom: 10px;
+  }
+  
+  .chatbot-send-button {
+    border-radius: 5px;
+  }
+
