--- mindfullcampus/src/components/home.css
@@ -1,68 +1,0 @@
-.formContainer {
-    /* Existing background styles */
-    /* background-color: #000000; */
-    height: 100vh;
-  
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-image: url("data:image/svg+xml,%3Csvg width='64' height='64' viewBox='0 0 64 64' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M8 16c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zm33.414-6l5.95-5.95L45.95.636 40 6.586 34.05.636 32.636 2.05 38.586 8l-5.95 5.95 1.414 1.414L40 9.414l5.95 5.95 1.414-1.414L41.414 8zM40 48c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zM9.414 40l5.95-5.95-1.414-1.414L8 38.586l-5.95-5.95L.636 34.05 6.586 40l-5.95 5.95 1.414 1.414L8 41.414l5.95 5.95 1.414-1.414L9.414 40z' fill='%23ffffff' fill-opacity='0.35' fill-rule='evenodd'/%3E%3C/svg%3E");
-    
-}
-
-.logout {
-    position: absolute;
-    top: 90px;
-    left: 80%;
-}
-
-.pageTitle {
-    text-align: center;
-    /* Replace with a suitable color that matches the background */
-    ;
-    /* Adjust the size as needed */
-    margin-top: 20px;
-    /* Add margin to position it correctly */
-    position: absolute;
-    top: 20px;
-    left: 50%;
-    transform: translateX(-50%);
-
-}
-
-.pageTitle,
-h1 {
-
-    font-size: 60px;
-    /* Adjust the size as needed */
-    color: #f9f8fa;
-    /* Replace with a suitable color that matches the background */
-    text-align: center;
-    padding: 20px;
-    background-color: #360c96;
-    border-radius: 10px;
-    margin-top: 40px;
-    /* Add margin to position it correctly */
-
-
-}
-
-
-
-
-.card:hover {
-    transform: translateY(-5px);
-}
-
-
-.card-row {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 20px;
-  gap: 125px;
-  
-}
-
-.litxt{
-  text-decoration: none;
-}
