+++ mindfullcampus/src/components/images/.~lock.M.STOCK PPT.pptx#
@@ -1,0 +1,1 @@
+,lex,lex,28.06.2023 19:22,file:///home/lex/.config/libreoffice/4;
