+++ mindfullcampus/src/components/menstrual/menstrual.css
@@ -1,0 +1,117 @@
+.menstrual-page {
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  
+  header h1 {
+    background-color: white;
+    color: #333;
+    margin-bottom: 20px;
+  }
+  
+  section {
+    margin-bottom: 40px;
+  }
+  
+  h2 {
+    color: #333;
+    font-size: 24px;
+    margin-bottom: 10px;
+  }
+  
+  p {
+    margin-bottom: 10px;
+  }
+  
+  section ul.symptom-list {
+    list-style-type: none;
+    padding-left: 0;
+    margin-bottom: 10px;
+    list-style : none;
+    
+  }
+  
+  section ul.symptom-list li {
+    margin-bottom: 10px;
+    list-style :none;
+  }
+  
+  section ul.symptom-list li::before {
+    content: "•";
+    color: #333;
+    font-size: 14px;
+    margin-right: 5px;
+  }
+  
+  section ul.symptom-list li .symptom {
+    font-weight: bold;
+  }
+  
+  section ul.symptom-list li .tip {
+    margin-left: 10px;
+  }
+
+  
+  
+   
+  .product-card-top{
+    display: flex;
+    flex-direction: row;
+    padding-top: 55px;
+  }
+  
+  
+  
+  
+  .product-card {
+    padding-top: 155px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-bottom: 20px;
+    padding: 20px;
+    background-color: #f5f5f5;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .product-card img {
+    width: 300px;
+    height: 300px;
+    margin-bottom: 10px;
+  }
+  
+  .product-card h3 {
+    color: #333;
+    font-size: 20px;
+    margin-bottom: 10px;
+  }
+  
+  footer {
+    text-align: center;
+    color: #666;
+    margin-top: 40px;
+  }
+
+  .topimg {
+    max-width: 100px;
+    
+    height: auto;
+    margin-bottom: 10px;
+  }
+
+  .toppart img{
+    width: 100%;
+    height: auto;
+  }
+
+.toppart-heading{
+    padding-top: 25px;
+}
+  
+.toppart-para{
+    padding-top: 20px;
+    padding-bottom: 25px;
+}
+
