--- mindfullcampus/src/components/physicalhealth.css
@@ -1,92 +1,0 @@
-
-
-.formContainer {
-
- 
-    background-color: #f2f0f5;
-    height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  
-    background-color: #360c96;
-    background-image: url("data:image/svg+xml,%3Csvg width='64' height='64' viewBox='0 0 64 64' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M8 16c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zm33.414-6l5.95-5.95L45.95.636 40 6.586 34.05.636 32.636 2.05 38.586 8l-5.95 5.95 1.414 1.414L40 9.414l5.95 5.95 1.414-1.414L41.414 8zM40 48c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zM9.414 40l5.95-5.95-1.414-1.414L8 38.586l-5.95-5.95L.636 34.05 6.586 40l-5.95 5.95 1.414 1.414L8 41.414l5.95 5.95 1.414-1.414L9.414 40z' fill='%23ffffff' fill-opacity='0.35' fill-rule='evenodd'/%3E%3C/svg%3E");
-}
-.formWrapper {
-      width: 580px;
-      background-color: white;
-      padding: 100px 60px;
-      border-radius: 10px;
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-      align-items: center;
-      
-    } 
-    .formWrapper h1{
-    font-size: medium;
-    color: azure;
-    }
-      .logo  {
-    font-weight:bolder;
-    font-size:10px;
-    color:#f0f0f1;
-    padding-bottom: 30px;
-        
-      }
-  
-      .title {
-        padding-top: 10px;
-        
-        color: #f8f8f8;
-        font-size: 21px;
-      }
-  
-      form {
-       
-        display: flex;
-        flex-direction: column;
-        gap: 15px;
-      }
-        input {
-          padding: 15px;
-          border: none;
-          width: 400px;
-          border-bottom: 1px solid #360c96;
-        }
-        select{
-            padding: 15px;
-            border: none;
-            width: 400px;
-            border-bottom: 1px solid #360c96;
-        }
-    
-        button {
-          background-color: #360c96;
-          color: white;
-          
-          font-weight: bold;
-          border: none;
-          cursor: pointer;
-          border-radius: 5px;
-        }
-  
-        label {
-          display: flex;
-          align-items: center;
-          gap: 10px;
-          color: #db4a26;
-          font-size: 12px;
-          cursor: pointer;
-        }
-          img {
-            width: 32px;
-          }
-        
-      
-      p {
-        color: #5d5b8d;
-        font-size: 15px;
-        margin-top: 10px;
-      
-  }
